+++ R/helper_functions.R
@@ -229,7 +229,7 @@
     "extra", "scale", "marginal", "alpha", "beta", "survival", "infrequent_purchase",
     "auxiliary", "mix", "shiftprop", "phi", "ndt", "hu", "xi", "coi", "zoi",
     "aux", "dist", "selection", "outcome", "time_dummies", "sigma_random",
-    "beta_random", "car", "nominal"
+    "beta_random", "car", "nominal", "bidrange"
   )
 }